+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Open to nested action via `CommandPalette.of(context).openToAction(actionId)`
+### Changed
+- When a nested action is selected, the label of that action will prefix the command palette text field. This can be enabled by setting `prefixNestedActions` to true (this is also the current default) in `CommandPaletteStyle`
 
 ### Fixed
 - BREAKING: default open key is now platform dependent. Previously it was always Ctrl+K, but now it will check if the platform is MacOS (this includes Web running on Mac) and if so will change the default open key to Command. While this change does make things function as I originally intended, this is changing default behavior so I'm considering this a breaking change